feat(api): port remaining 11 routes to app router handlers; pages/api retired
automation, bookmarks, dataset assets, ga, mailchimp, newsletter, packages,
projects, github stats, and the view counters — identical logic, shapes, and
status codes. Supabase client construction moves to request time in the view
routes (module scope would run during build page-data collection).
